+++ apps/client/src/plugins/Fir/FirToolbarView.js
@@ -27,7 +27,8 @@ const StyledButtonGroup = styled(ButtonGroup)(({ theme }) => ({
   whiteSpace: "nowrap",
 }));
 
-const IconButton = styled(Button)(({ theme }) => ({
+const IconButton = styled(Button)(({ theme, on, invert }) => ({
+  ...(on === "false" ? { backgroundColor: "#dcdcdc" } : {}),
   margin: theme.spacing(0),
   paddingLeft: 0,
   paddingRight: 0,
@@ -37,14 +38,6 @@ const IconButton = styled(Button)(({ theme }) => ({
   "& img": {
     filter: on === "true" && invert === "true" ? "invert(1)" : "", // fixes icon-colors on geometry icons.
   },
-  variants: [
-    {
-      props: {
-        on: "false",
-      },
-      style: { backgroundColor: "#dcdcdc" },
-    },
-  ],
 }));
 
 const FileInputContainer = styled("div")(({ theme }) => ({